requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
langchain-core>=0.1.0
langgraph>=0.0.40
rapidfuzz>=3.6.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
httpx>=0.25.0  # for mocking platform API calls
pytest>=7.4.0
pytest-asyncio>=0.21.0
